obtain sample attributes for a study.

obtain sample attributes for a study.

samp_atts_by_study(srp = "SRP082656", size = 500, cursor = NULL, ...)

Arguments

srp
character(1) SRP tag for a study in NCBI SRA

size
numeric(1) maximum number of records to return

cursor
character(1) the value of the cursor returned in a previous call; see note.

...
not currently used

Note

If the number of records available in query response exceeds the value of size, the returned tibble will have an attribute 'cursor' which can be retrieved and used in subsequent calls to this function.

If size is zero, a list is returned with information on the number of records available.

Examples

t1 = samp_atts_by_study(size=10)
t1

#> # A tibble: 10 x 11
#>    accession collection_date env_biome env_feature env_material geo_loc_name
#>    <chr>     <chr>           <chr>     <chr>       <chr>        <chr>       
#>  1 SRR40520… 2014            Human st… Human-asso… Human infan… Italy: Tren…
#>  2 SRR40520… 2014            Human st… Human-asso… Human infan… Italy: Tren…
#>  3 SRR40520… 2014            Human st… Human-asso… Human mothe… Italy: Tren…
#>  4 SRR40520… 2014            Human st… Human-asso… Human mothe… Italy: Tren…
#>  5 SRR40520… 2015            Human st… Human-asso… Human mothe… Italy: Tren…
#>  6 SRR40520… 2014            Human st… Human-asso… Human mothe… Italy: Tren…
#>  7 SRR40520… 2014            Human st… Human-asso… Human mothe… Italy: Tren…
#>  8 SRR40520… 2015            Human st… Human-asso… Human mothe… Italy: Tren…
#>  9 SRR40520… 2014            Human mi… Human-asso… Human mothe… Italy: Tren…
#> 10 SRR40520… 2014            Human mi… Human-asso… Human mothe… Italy: Tren…
#> # … with 5 more variables: host <chr>, lat_lon <chr>, timepoint <chr>,
#> #   gender <chr>, couple <chr>

attr(t1, "cursor")

#> [1] "X2lkfHxhc2N8fHx8U1JSNDA1MjAzMA=="

samp_atts_by_study(size=0)

#> $hits
#> list()
#> 
#> $cursor
#> NULL
#> 
#> $stats
#> $stats$total
#> [1] 28
#> 
#> $stats$took
#> [1] 0
#> 
#> 
#> $success
#> [1] TRUE
#>